[RUNTIEM] project structure for C API
