+++ include/grpc_c/grpc_c.h
@@ -32,12 +32,25 @@
  */
 
 
-#ifndef TEST_GRPC_C_UNARY_BLOCKING_CALL_PUBLIC_H
-#define TEST_GRPC_C_UNARY_BLOCKING_CALL_PUBLIC_H
+#ifndef GRPC_C_PUBLIC_H
+#define GRPC_C_PUBLIC_H
 
-#include "grpc_c_public.h"
+#include <stdlib.h>
+#include <grpc_c/message.h>
 
-GRPC_status GRPC_unary_blocking_call(GRPC_channel *channel, const GRPC_method *const rpc_method,
-                                     GRPC_context *const context, const GRPC_message message, GRPC_message *response);
+typedef struct grpc_channel GRPC_channel;
+typedef struct grpc_status GRPC_status;
+typedef struct grpc_context GRPC_context;
 
-#endif //TEST_GRPC_C_UNARY_BLOCKING_CALL_PUBLIC_H
+typedef struct GRPC_method {
+  enum RpcType {
+    NORMAL_RPC = 0,
+    CLIENT_STREAMING,  // request streaming
+    SERVER_STREAMING,  // response streaming
+    BIDI_STREAMING
+  } type;
+  const char* name;
+} GRPC_method;
+
+
+#endif // GRPC_C_PUBLIC_H
